Fix possible issue with logging for Linq Readonly tests
